# docs/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "学科知识整合智能体",
    "description": "学科知识整合智能体 API：多格式解析、统一 JSON、证据链和任务状态接口。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/textbooks": {
      "get": {
        "tags": [
          "textbooks"
        ],
        "summary": "List Textbooks",
        "operationId": "list_textbooks_api_textbooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextbookListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/textbooks/upload": {
      "post": {
        "tags": [
          "textbooks"
        ],
        "summary": "Upload Textbook",
        "operationId": "upload_textbook_api_textbooks_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_textbook_api_textbooks_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TextbookUploadResponse"
                    },
                    {
                      "$ref": "#/components/schemas/TextbookBatchUploadResponse"
                    }
                  ],
                  "title": "Response Upload Textbook Api Textbooks Upload Post"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/textbooks/upload-batch": {
      "post": {
        "tags": [
          "textbooks"
        ],
        "summary": "Upload Textbooks Batch",
        "operationId": "upload_textbooks_batch_api_textbooks_upload_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_textbooks_batch_api_textbooks_upload_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextbookBatchUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/textbooks/upload-async": {
      "post": {
        "tags": [
          "textbooks"
        ],
        "summary": "Upload Textbook Async",
        "operationId": "upload_textbook_async_api_textbooks_upload_async_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_textbook_async_api_textbooks_upload_async_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncTextbookParseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/textbooks/{raw_file_id}/parse": {
      "post": {
        "tags": [
          "textbooks"
        ],
        "summary": "Parse Textbook",
        "operationId": "parse_textbook_api_textbooks__raw_file_id__parse_post",
        "parameters": [
          {
            "name": "raw_file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Raw File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextbookUploadResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/textbooks/{raw_file_id}/parse-async": {
      "post": {
        "tags": [
          "textbooks"
        ],
        "summary": "Parse Textbook Async",
        "operationId": "parse_textbook_async_api_textbooks__raw_file_id__parse_async_post",
        "parameters": [
          {
            "name": "raw_file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Raw File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncTextbookParseResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/textbooks/{raw_file_id}": {
      "get": {
        "tags": [
          "textbooks"
        ],
        "summary": "Get Textbook",
        "operationId": "get_textbook_api_textbooks__raw_file_id__get",
        "parameters": [
          {
            "name": "raw_file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Raw File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedTextbook"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "title": "ApiErrorResponse"
      },
      "AsyncTextbookParseResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRecord"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "default": true
          },
          "upload_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job"
        ],
        "title": "AsyncTextbookParseResponse"
      },
...
      "Body_upload_textbook_api_textbooks_upload_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "textbook_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textbook Title"
          }
        },
        "type": "object",
        "title": "Body_upload_textbook_api_textbooks_upload_post"
      },
      "Body_upload_textbook_async_api_textbooks_upload_async_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "textbook_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textbook Title"
          }
        },
        "type": "object",
        "title": "Body_upload_textbook_async_api_textbooks_upload_async_post"
      },
      "Body_upload_textbooks_batch_api_textbooks_upload_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "textbook_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textbook Title"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_textbooks_batch_api_textbooks_upload_batch_post"
      },
...
      "JobRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PipelineStepRecord"
            },
            "type": "array",
            "title": "Steps"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "context_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status"
        ],
        "title": "JobRecord"
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "textbook_upload",
          "textbook_batch_upload",
          "textbook_parse",
          "textbook_pipeline",
          "large_file_upload",
          "graph_build",
          "converted_textbook_import"
        ],
        "title": "JobType"
      },
...
      "ParsedTextbook": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "raw_file": {
            "$ref": "#/components/schemas/RawFile"
          },
          "elements": {
            "items": {
              "$ref": "#/components/schemas/DocumentElement"
            },
            "type": "array",
            "title": "Elements"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/Chunk"
            },
            "type": "array",
            "title": "Chunks"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "raw_file",
          "elements",
          "sections",
          "chunks"
        ],
        "title": "ParsedTextbook"
      },
...
      "TextbookBatchUploadResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRecord"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TextbookUploadResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/TextbookUploadError"
            },
            "type": "array",
            "title": "Errors"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "success_count": {
            "type": "integer",
            "title": "Success Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job",
          "items",
          "total_count",
          "success_count",
          "failed_count"
        ],
        "title": "TextbookBatchUploadResponse"
      },
      "TextbookListResponse": {
        "properties": {
          "textbooks": {
            "items": {
              "$ref": "#/components/schemas/TextbookSummary"
            },
            "type": "array",
            "title": "Textbooks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "textbooks"
        ],
        "title": "TextbookListResponse"
      },
      "TextbookSummary": {
        "properties": {
          "raw_file_id": {
            "type": "string",
            "title": "Raw File Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "element_count": {
            "type": "integer",
            "title": "Element Count"
          },
          "section_count": {
            "type": "integer",
            "title": "Section Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "parsed_output_path": {
            "type": "string",
            "title": "Parsed Output Path"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "raw_file_id",
          "title",
          "format",
          "element_count",
          "section_count",
          "chunk_count",
          "parsed_output_path",
          "updated_at"
        ],
        "title": "TextbookSummary"
      },
...
      "TextbookUploadResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRecord"
          },
          "raw_file_id": {
            "type": "string",
            "title": "Raw File Id"
          },
          "parsed_output_path": {
            "type": "string",
            "title": "Parsed Output Path"
          },
          "parsed_textbook": {
            "$ref": "#/components/schemas/ParsedTextbook"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job",
          "raw_file_id",
          "parsed_output_path",
          "parsed_textbook"
        ],
        "title": "TextbookUploadResponse"
      },
...
    }
  }
}